create calendar.js to separate logic from CalendarBoard component
